+++ swak/__init__.py
@@ -1,0 +1,1 @@
+# ToDo: Make & use Literal types for Enums!